boards/arm/stm32f7: switch board common to boards/arm/common/stm32
switch stm32f7 board common to boards/arm/common/stm32.

Move the STM32F7-only board-common features into the shared directory
so they are available to every STM32 family:

  - bmi270, cs4344, can_setup, cansock_setup, reset, romfs_initialize and
    spitest
  - the duplicate stm32_bh1750 is removed (already present in common).
