--- boards/arm/stm32f7/common/src/Make.defs
@@ -1,62 +1,0 @@
-#############################################################################
-# boards/arm/stm32f7/common/src/Make.defs
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.  The
-# ASF licenses this file to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance with the
-# License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-# License for the specific language governing permissions and limitations
-# under the License.
-#
-#############################################################################
-
-ifeq ($(CONFIG_ARCH_BOARD_COMMON),y)
-
-ifeq ($(CONFIG_SENSORS_BH1750FVI),y)
-  CSRCS += stm32_bh1750.c
-endif
-
-ifeq ($(CONFIG_SENSORS_BMI270_I2C),y)
-  CSRCS += stm32_bmi270.c
-endif
-
-ifeq ($(CONFIG_AUDIO_CS4344),y)
-  CSRCS += stm32_cs4344.c
-endif
-
-ifeq ($(CONFIG_STM32_CAN),y)
-ifeq ($(CONFIG_STM32_CAN_CHARDRIVER),y)
-CSRCS += stm32_can_setup.c
-endif
-ifeq ($(CONFIG_STM32_CAN_SOCKET),y)
-CSRCS += stm32_cansock_setup.c
-endif
-endif
-
-ifeq ($(CONFIG_STM32_ROMFS),y)
-CSRCS += stm32_romfs_initialize.c
-endif
-
-ifeq ($(CONFIG_BOARDCTL_RESET),y)
-CSRCS += stm32_reset.c
-endif
-
-ifeq ($(CONFIG_STM32_SPI_TEST),y)
-CSRCS += stm32_spitest.c
-endif
-
-DEPPATH += --dep-path src
-VPATH += :src
-CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)src
-
-endif
